+++ src/BrandId.php
@@ -3,7 +3,7 @@
  * Brand ID
  *
  * @author Pronamic <info@pronamic.eu>
- * @copyright 2005-2020 Pronamic
+ * @copyright 2005-2021 Pronamic
  * @license GPL-3.0-or-later
  * @package Pronamic\WordPress\Pay\Gateways\Payvision
  */
@@ -17,7 +17,7 @@ use Pronamic\WordPress\Pay\Core\PaymentMethods;
  *
  * @link https://developers.acehubpaymentservices.com/v3.3/reference#brands
  * @author Remco Tolsma
- * @version 1.0.0
+ * @version 1.1.0
  * @since 1.0.0
  */
 class BrandId {