+++ src/TrackingCode.php
@@ -3,7 +3,7 @@
  * Tracking Code
  *
  * @author    Pronamic <info@pronamic.eu>
- * @copyright 2005-2020 Pronamic
+ * @copyright 2005-2021 Pronamic
  * @license   GPL-3.0-or-later
  * @package   Pronamic\WordPress\Pay\Gateways\Payvision
  */
@@ -15,7 +15,7 @@ namespace Pronamic\WordPress\Pay\Gateways\Payvision;
  *
  * @link    https://developers.acehubpaymentservices.com/v3.3/reference#payment-3-1
  * @author  Remco Tolsma
- * @version 1.0.0
+ * @version 1.1.0
  * @since   1.0.0
  */
 class TrackingCode implements \JsonSerializable {