+++ tests/src/GatewayTest.php
@@ -3,7 +3,7 @@
  * Gateway Test
  *
  * @author    Pronamic <info@pronamic.eu>
- * @copyright 2005-2020 Pronamic
+ * @copyright 2005-2021 Pronamic
  * @license   GPL-3.0-or-later
  * @package   Pronamic\WordPress\Pay\Gateways\Payvision
  */
@@ -19,7 +19,7 @@ use Pronamic\WordPress\Pay\Payments\PaymentStatus;
  * Gateway Test
  *
  * @author  Remco Tolsma
- * @version 1.0.0
+ * @version 1.1.0
  * @since   1.0.0
  */
 class GatewayTest extends \WP_UnitTestCase {