+++ CHANGELOG.md
@@ -7,11 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased][unreleased]
 -
 
+## [1.1.0] - 2021-02-08
+- Added transaction description.
+- Added advanced purchase ID setting.
+
 ## [1.0.1] - 2021-01-18
 - Added business ID to gateway ID column in payments overview.
 
 ## 1.0.0 - 2020-11-18
 - First release.
 
-[unreleased]: https://github.com/wp-pay-gateways/adyen/compare/1.0.1...HEAD
+[unreleased]: https://github.com/wp-pay-gateways/adyen/compare/1.1.0...HEAD
+[1.1.0]: https://github.com/wp-pay-gateways/payvision/compare/1.0.1...1.1.0
 [1.0.1]: https://github.com/wp-pay-gateways/adyen/compare/1.0.0...1.0.1